refactor so that GenericEvent and Header classes are in same package as other classes
