+++ src/main/java/uk/gov/ons/ctp/integration/rhsvc/message/impl/CaseEvent.java
@@ -4,7 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import uk.gov.ons.ctp.integration.rhsvc.message.GenericEvent;
 
 @Data
 @Builder
